refactor(asset): simplify folder structure into model, asset, output and system
This commit does not change the model structure as much. The only major change in model folder is
adding back the `bsdf` folder. BSDF files belong to model folder and are reusable for several
studies.

All the other input files are grouped into an `asset` folder. Asset folder is not part
of the geometry or geometry metadata but is needed to run a simulation. Skies, sensor-grids, views
and IES files are all located in asset folder. Among these files I wonder if IES should be moved to
model folder as it is basically a definition for electric lights similar to what BSDF is for
defining materials.

output folder will include all different types of outputs and system folder
will include settings for runs. system folder should be where data like options for radiance
commands for a study is saved.

BREAKING CHANGE: The changes in folder structure will break current scripts that rely on older
folder sturcture.

no open issues actually! oops.